Updated WatcherService.cs to allow for creation of cleaned .log files instead of .txt

+++ Cleaner/WatcherService.cs
@@ -1,5 +1,7 @@
-﻿﻿using System.IO;
+﻿using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Text.RegularExpressions;
 using Serilog.Events;
 
 namespace Cleaner
@@ -8,10 +10,12 @@ namespace Cleaner
     {
         private readonly FileSystemWatcher _watcher;
         private readonly CliOptions _options;
+        private readonly IEnumerable<Regex> _regexFilters;
 
         public WatcherService(CliOptions options)
         {
             _options = options;
+            _regexFilters = _options.GetRegexFilters();
             var directoryInfo = new DirectoryInfo(_options.Directory);
             _watcher = new FileSystemWatcher(directoryInfo.FullName, _options.Pattern)
             {
@@ -37,18 +41,37 @@ namespace Cleaner
 
         private void OnFileChanged(object sender, FileSystemEventArgs e)
         {
+            if (IsFileAlreadyFiltered(e.Name))
+            {
+                Logger.Log(
+                    "Detected file change in already filtered file: {file}. This event is being ignored.",
+                    LogEventLevel.Verbose,
+                    e.FullPath);
+                return;
+            }
+
             Logger.Log("Detected file change: {file}", LogEventLevel.Debug, e.FullPath);
             CleanFile(e.FullPath);
         }
 
         private void OnFileCreated(object sender, FileSystemEventArgs e)
         {
+            if (IsFileAlreadyFiltered(e.Name))
+            {
+                Logger.Log(
+                    "Detected file creation in already filtered file: {file}. This event is being ignored.",
+                    LogEventLevel.Verbose,
+                    e.FullPath);
+                return;
+            }
+
             Logger.Log("Detected file creation: {file}", LogEventLevel.Debug, e.FullPath);
             CleanFile(e.FullPath);
         }
 
         private void CleanFile(string fileName)
         {
+            if (IsFileAlreadyFiltered(fileName)) return;
             var tempFileName = $"{fileName}.temp";
             if (File.Exists(tempFileName)) File.Delete(tempFileName);
             File.Copy(fileName, tempFileName);
@@ -57,10 +80,8 @@ namespace Cleaner
                 var content = File.ReadLines(tempFileName).ToList();
                 var filteredContent = content
                     .Where(line => _options.KeepRows
-                        ? _options.GetRegexFilters()
-                            .Any(regex => regex.IsMatch(line))
-                        : _options.GetRegexFilters()
-                            .All(regex => !regex.IsMatch(line)))
+                        ? _regexFilters.Any(regex => regex.IsMatch(line))
+                        : _regexFilters.All(regex => !regex.IsMatch(line)))
                     .ToList();
                 var removedLines = content.Count - filteredContent.Count;
                 Logger.Log("Filtered {filterCount} out of {@overallCount} lines",
@@ -68,9 +89,10 @@ namespace Cleaner
                     removedLines,
                     content.Count);
                 var fileNameWithoutExtension = Path.GetFileNameWithoutExtension(fileName);
+                var fileExtension = Path.GetExtension(fileName);
                 var filterDirectory = new DirectoryInfo(_options.FilterDirectory);
                 if (!filterDirectory.Exists) Directory.CreateDirectory(filterDirectory.FullName);
-                var filterFile = new FileInfo($"{fileNameWithoutExtension}{_options.Suffix}.txt");
+                var filterFile = new FileInfo($"{fileNameWithoutExtension}{_options.Suffix}{fileExtension}");
                 var combinedFilterFile = Path.Combine(filterDirectory.FullName, filterFile.Name);
                 File.WriteAllLines(combinedFilterFile, filteredContent);
                 Logger.Log("Filtered {removedLines} lines from {file}.", LogEventLevel.Information, removedLines,
@@ -86,6 +108,13 @@ namespace Cleaner
         public void StartWatching()
         {
             _watcher.EnableRaisingEvents = true;
+        }
+
+        private bool IsFileAlreadyFiltered(string fileName)
+        {
+            var fileNameWithoutExtension = Path.GetFileNameWithoutExtension(fileName);
+            return fileNameWithoutExtension != null &&
+                   fileNameWithoutExtension.EndsWith(_options.Suffix);
         }
     }
 }
